chico_delivery/ext/cli/__init__.py
import click
from chico_delivery.ext.db import db
from chico_delivery.ext.db import models #noqa


def init_app(app):

    @app.cli.command()
    def create_db():
        """Este comando inicializa o db"""
        db.create_all()


    @app.cli.command()
    def listar_pedidos():
        # TODO: usar tabulate
        click.echo("lista de pedidos")